+++ src/main/java/panacea/website_dat_lich_khach_san/repository/HotelRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import panacea.website_dat_lich_khach_san.entity.BookingDetail;
 @Repository
-public interface HotelRepository extends JpaRepository<BookingDetail, Long> {
+public interface HotelRepository extends JpaRepository<BookingDetail, Integer> {
 }
